Move SnapshotId to config-provisioning
