+++ config-provisioning/src/main/java/com/yahoo/config/provision/SnapshotId.java
@@ -1,4 +1,4 @@
-package com.yahoo.vespa.hosted.provision.backup;
+package com.yahoo.config.provision;
 
 import java.util.Objects;
 import java.util.UUID;
